% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest2.R
\name{design_report}
\alias{design_report}
\title{Title}
\usage{
design_report(ctx, obj_name = "t", stringify = TRUE)
}
\arguments{
\item{ctx}{}

\item{obj_name}{}

\item{stringify}{}
}
\value{

}
\description{
Title
}
