% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest2.R
\name{test}
\alias{test}
\title{Title}
\usage{
test(
  ctx,
  alpha = NULL,
  side = NULL,
  es0 = NULL,
  obj_name = "t",
  stringify = TRUE
)
}
\arguments{
\item{ctx}{}

\item{alpha}{}

\item{side}{}

\item{es0}{}

\item{obj_name}{}

\item{stringify}{}
}
\value{

}
\description{
Title
}
